+++ clove/network/bitcoin/__init__.py
@@ -1,8 +1,13 @@
-from clove.block_explorer.blockcypher import BlockcypherAPI
+from typing import Optional
+
+from clove.block_explorer.insight import InsightAPIv4
 from clove.network.bitcoin.base import BitcoinBaseNetwork
+from clove.utils.bitcoin import from_base_units
+from clove.utils.external_source import clove_req_json
+from clove.utils.logging import logger
 
 
-class Bitcoin(BlockcypherAPI, BitcoinBaseNetwork):
+class Bitcoin(InsightAPIv4, BitcoinBaseNetwork):
     """
     Class with all the necessary BTC network information based on
     https://github.com/bitcoin/bitcoin/blob/master/src/chainparams.cpp
@@ -26,7 +31,19 @@ class Bitcoin(BlockcypherAPI, BitcoinBaseNetwork):
         'SECRET_KEY': 128
     }
     source_code_url = 'https://github.com/bitcoin/bitcoin/blob/master/src/chainparams.cpp'
-    blockexplorer_tx = 'https://live.blockcypher.com/btc/tx/{0}/'
+    api_url = 'https://insight.bitpay.com/api'
+    ui_url = 'https://insight.bitpay.com'
+    fee_endpoint = 'https://api.blockcypher.com/v1/btc/main'
+
+    @classmethod
+    def get_fee(cls) -> Optional[float]:
+        '''Returns actual fee per kb.'''
+        response = clove_req_json(cls.fee_endpoint)
+        fee = response.get('high_fee_per_kb')
+        if not fee:
+            logger.error('Cannot find the right key (high_fee_per_kb) while getting fee in blockcypher.')
+            return
+        return from_base_units(fee)
 
 
 class BitcoinTestNet(Bitcoin):
@@ -50,4 +67,6 @@ class BitcoinTestNet(Bitcoin):
         'SECRET_KEY': 239
     }
     testnet = True
-    blockexplorer_tx = 'https://live.blockcypher.com/btc-testnet/tx/{0}/'
+    api_url = 'https://test-insight.bitpay.com/api'
+    ui_url = 'https://test-insight.bitpay.com'
+    fee_endpoint = 'https://api.blockcypher.com/v1/btc/test3'